Rewrite theme using git-info
This uses the verbose mode to be able to display indexed, unindexed and
untracked files indicators. This is the last theme left to be migrated
to git-info, and the first one to use its verbose mode.

Originally, the theme also supported svn via vcs_info. The svn support
is lost in this rewrite. Maybe if a new svn-info module is introduced to
Zim...

Closes zimfw/zimfw#157

Copied from zimfw/zimfw@f3e673a52b5f3e2c1cd8e54fc55640166b4d9a57